+++ scripts/sdxl_peft.py
@@ -329,7 +329,6 @@ def train(cfg: SDXLPeftConfig, strategies: "SdxlPeftStrategy"):
         optimizer_train_fn,
         optimizer_eval_fn,
         lr_descriptions,
-        text_encoder_lr  # TODO: why only text_encoder_lr here?
     ) = prepare_optimizer(cfg.optimizer, cfg.peft, adapter)
 
     # prepare dataloader
@@ -535,7 +534,6 @@ def train(cfg: SDXLPeftConfig, strategies: "SdxlPeftStrategy"):
         num_train_epochs=num_train_epochs,
         optimizer_name=optimizer_name,
         optimizer_args=optimizer_args,
-        text_encoder_lr=text_encoder_lr,  # TODO: why only text_encoder_lr here?
         net_kwargs=net_kwargs,
         train_dataloader=train_dataloader,
         total_batch_size=total_batch_size,